ya jala esta cochinada

+++ src/services/productService.js
@@ -1,6 +1,6 @@
-
 const Product = require('../models/productModel');
 const Brand = require('../models/brandModel');
+const facturapi = require('../apis/facturapi');
 
 module.exports = {
     getAllProducts: async () => {
@@ -9,23 +9,37 @@ module.exports = {
 
     createProduct: async ({ name, desc, price, category, brandId, quantity, images }) => {
         const product = new Product({ name, desc, price, category, brand: brandId, quantity, images });
+        const facturapiProduct = await facturapi.createProduct(product);
+
+        product.facturapi = facturapiProduct.id;
+        console.log(facturapiProduct);
+        console.log(product);
+        
         return await product.save();
     },
 
     updateProduct: async (_id, updates) => {
-        const updatedProduct = await Product.findByIdAndUpdate(_id, updates, { new: true }).populate('brand');
-        if (!updatedProduct) {
-            throw new Error(`Product with ID: ${_id} not found.`);
-        }
-        return updatedProduct;
+        const product = await Product.findById(_id);
+        if (!product) throw new Error(`Product with ID: ${_id} not found.`);
+
+        const facturapiData = {
+            description: updates.desc || product.desc,
+            price: updates.price || product.price,
+            product_key: "50202306" 
+        };
+
+        await facturapi.updateProduct(product.facturapi, facturapiData);
+        return await Product.findByIdAndUpdate(_id, updates, { new: true });
     },
 
     deleteProduct: async (_id) => {
-        const deletedProduct = await Product.findByIdAndDelete(_id);
-        if (!deletedProduct) {
-            throw new Error(`Product with ID: ${_id} not found.`);
-        }
-        return deletedProduct;
+        const product = await Product.findById(_id);
+        if (!product) throw new Error(`Product with ID: ${_id} not found.`);
+
+        const productDeleted = await facturapi.deleteClient(product.facturapi);
+        if (!productDeleted) throw new Error(`Product with ID: ${_id} couldn't be deleted from Facturapi.`);
+
+        return await Product.findByIdAndDelete(_id);
     },
 
     getBrandById: async (brandId) => {
