+++ src/services/userService.js
@@ -1,4 +1,5 @@
 const User = require('../models/userModel');
+const facturapi = require('../apis/facturapi');
 
 module.exports = {
     getAllUsers: async () => {
@@ -10,19 +11,36 @@ module.exports = {
     },
 
     createUser: async (input) => {
-        const user = new User(input);
+        const facturapiClient = await facturapi.createClient(input);
+        const user = new User({
+            ...input,
+            facturapi: facturapiClient.id
+        });
+
         return await user.save();
     },
 
     deleteUser: async (_id) => {
-        const user = await User.findByIdAndDelete(_id);
+        const user = await User.findById(_id);
         if (!user) throw new Error(`User with ID: ${_id} not found.`);
-        return user;
+
+        const clientDeleted = await facturapi.deleteClient(user.facturapi);
+        if (!clientDeleted) throw new Error(`User with ID: ${_id} couldn't be deleted from Facturapi.`);
+        
+        return await User.findByIdAndDelete(_id);
     },
 
     updateUser: async (_id, updates) => {
-        const user = await User.findByIdAndUpdate(_id, updates, { new: true });
+        const user = await User.findById(_id);
         if (!user) throw new Error(`User with ID: ${_id} not found.`);
-        return user;
+
+        const facturapiData = {
+            legal_name: updates.nombreCompleto || user.nombreCompleto,
+            email: updates.email || user.email,
+            address: { zip: updates.direccion || user.direccion }
+        };
+
+                await facturapi.updateClient(user.facturapi, facturapiData);
+        return  await User.findByIdAndUpdate(_id, updates, { new: true });
     }
 };
